#252 Feat: uuid 위치 조정 및 팀 이미지 관련 코드 수정
