+++ src/main/java/com/codiary/backend/global/common/uuid/Uuid.java
@@ -1,0 +1,28 @@
+package com.codiary.backend.global.common.uuid;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Uuid {
+
+  @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Column(name = "uuid", columnDefinition = "varchar(256)")
+  private String uuid;
+
+  @Builder
+  public Uuid(String uuid) {
+    this.uuid = uuid;
+  }
+}
