+++ src/main/java/com/codiary/backend/global/s3/AmazonS3Manager.java
@@ -3,14 +3,13 @@ package com.codiary.backend.global.s3;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.codiary.backend.global.common.uuid.Uuid;
 import com.codiary.backend.global.config.S3Config;
-import com.codiary.backend.domain.member.entity.Uuid;
+import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
 
 @Slf4j
 @Component
